+++ examples/log_all_selene_commands_with_wait__framework/framework/__init__.py
@@ -1,0 +1,10 @@
+"""
+«framework» is kind of name for you project,
+so rename it as needed ;)
+
+E.g. if you write e2e tests for todomvc app,
+you can name your project as todomvc-e2e-test,
+and rename this framework py package to todomvc_e2e_test...
+"""
+
+from . import assist