+++ examples/log_all_selene_commands_with_wait__framework/framework/assist/python/logging.py
@@ -1,0 +1,20 @@
+import logging
+from functools import reduce
+from typing import Tuple, List
+
+
+class TranslatingFormatter(logging.Formatter):
+    translations: List[Tuple[str, str]] = ()
+
+    def formatMessage(self, record: logging.LogRecord) -> str:
+        original = super().formatMessage(record)
+
+        def translate(initial: str, item: Tuple[str, str]):
+            old, new = item
+            return initial.replace(old, new)
+
+        return reduce(
+            translate,
+            self.translations,
+            original,
+        )